requirements.txt
fastapi
uvicorn
gradio
faiss-cpu
sentence-transformers
pdfplumber
tqdm